Remove u-boot script and replace with extlinux.conf
The manual boot script specifies a fixed root device,
i.e., mmcblk0p2. However, how the kernel enumerates disks
is not deterministic, so relying on this can lead to boot
failures. It is better to rely on disk labels or UUIDs.

So, rely on wic's built-in way to create an extlinux.conf
file when assigning a UUID to the disk during creation.
This file specifies kernel and command line, so the randomly
generated UUID can be included there. U-Boot is instructed
to load extlinux.conf automatically by the default initial
environment.

Signed-off-by: Stefan Gloor <code@stefan-gloor.ch>