pya20: fix COMPATIBLE_MACHINE regex
The current regex in COMPATIBLE_MACHINE will also match machine
names which are not in the list, such as 'olinuxino-a20som204'.
Since 'olinuxino-a20som204' SOM is newer, a mapping.h file is not
available for it in upstream pya20, so there is no need to build
the recipe for it.

Add "^$" to the regex in order to restrict the pya20 recipe only to
those machines that have a mapping.h file.

Signed-off-by: Ovidiu Panait <ovidiu.panait.oss@gmail.com>